// src/app/models/DetalleComida.ts
export class DetalleComida {
    idComida: string;
    idAlimento: string;
    cantidad: number;
    observaciones: string;
    proteinasDetalleComida?: number;
    chsDetalleComida?: number;
    grasasDetalleComida?: number;
    kcalDetalleComida?: number;
    createdAt?: string;
    updatedAt?: string;
    _id?: string;
}